refactor: split the `scaffold.rs` file into a module
The `scaffold.rs` file was becoming too big. A module has been created,
populated by one file per scaffold sub-command.
